update pack tools

+++ PublicTools/bin/tools/PackTool/src/util/PackImg2Mng_PVR.py
@@ -16,6 +16,8 @@ import struct
 from pprint import pprint
 from struct import *
 
+from util import PackImgJPG
+
 def iter_find_files(path, fnexp):
     for root, dirs, files, in os.walk(path):
         for filename in fnmatch.filter(files, fnexp):
@@ -39,16 +41,7 @@ isSaveTransFile = False
 
 RGBMode = "PVR"
 
-isShowFilesCount = False
-GzipCnt = 0
-MngCnt = 0
-PkmCnt = 0
-PassCnt = 0
-
-filters = ["crater"]
-
-def work_file_PVR(filename, isAlphaJPG = False):
-    global GzipCnt, MngCnt, PkmCnt, PassCnt, filters
+def work_file_PVR(filename, isAlphaJPG = False, isTC4 = False):
     filepath = os.path.realpath(filename)
     filedir = os.path.dirname(filepath)
 
@@ -58,27 +51,6 @@ def work_file_PVR(filename, isAlphaJPG = False):
     preCMD = " -p "
     if not preAlpha:
         preCMD = ""
- 
-    for filtername in filters:
-        if filepath.find(filtername) != -1:
-            PassCnt = PassCnt + 1
-            passFiles.append(filename)
-            return
-    
-    with open(filepath, 'rb') as tmpFile:
-        tmpContent = tmpFile.read(3)
-        if tmpContent[0:2] == b'\x1f\x8b':
-            print ("Gzip File, pass.")
-            GzipCnt = GzipCnt + 1
-            return
-        if tmpContent == b"MNG":
-            print ("MNG File, pass.")
-            MngCnt = MngCnt + 1
-            return
-        if tmpContent == b"PKM":
-            print ("PKM File, pass.")
-            PkmCnt = PkmCnt + 1
-            return
             
     os.chdir(tpDir)
     
@@ -87,8 +59,9 @@ def work_file_PVR(filename, isAlphaJPG = False):
         isPng = True
     elif filename.find(".jpg") != -1:
         isPng = False
+        return PackImgJPG.convert(filepath)
     else:
-        return
+        return 2
 
     if isPng: 
         imgCmd = iosPngCmd           
@@ -96,15 +69,16 @@ def work_file_PVR(filename, isAlphaJPG = False):
         imgCmd = iosJpgCmd 
         
     if imgCmd == "":
-        PassCnt = PassCnt + 1
-        passFiles.append(filename)
-        return
+        return 2
        
     imgCmd = imgCmd % (filepath, filedir) 
     
-    rgbCMD = """ %s -f ETC1 %s -q etcslow -i %s -o %s """ % (pvrTexToolBin, preCMD, filepath, filepath.replace(".png", ".pvr"))
+    rgbCMD = """ %s -f PVRTC1_4_RGB %s -q pvrtcbest -i %s -o %s """ % (pvrTexToolBin, preCMD, filepath, filepath.replace(".png", ".pvr"))
     alphaCMD = """%s %s -alpha extract %s """ % (convertBin, filepath, filepath.replace(".png", ".alpha.jpg"))
-    alphaJPGCMD = """ %s -f ETC1 -q etcslow -i %s -o %s """ % (pvrTexToolBin, filepath.replace(".png", ".alpha.jpg"), filepath.replace(".png", ".alpha.pvr"))
+    alphaJPGCMD = """ %s -f PVRTC1_4_RGB -q pvrtcbest -i %s -o %s """ % (pvrTexToolBin, filepath.replace(".png", ".alpha.jpg"), filepath.replace(".png", ".alpha.pvr"))
+    
+    if isTC4:
+        rgbCMD = """ %s -f PVRTC1_4 %s -q pvrtcbest -i %s -o %s """ % (pvrTexToolBin, preCMD, filepath, filepath.replace(".png", ".pvr"))
     
     try:   
         if isPng:
@@ -114,32 +88,33 @@ def work_file_PVR(filename, isAlphaJPG = False):
                 os.remove(filepath.replace(".png", "_alpha.pkm"))
         
             os.system(rgbCMD) 
-            os.system(alphaCMD) 
             
-            if not isAlphaJPG:
+            if not isTC4:
+                os.system(alphaCMD) 
+            
+            if not isAlphaJPG and not isTC4:
                 os.system(alphaJPGCMD) 
-
+            
+            if not os.path.exists(filepath.replace(".png", ".pvr")):
+                return 2
+                
             os.rename(filepath.replace(".png", ".pvr"), filepath.replace(".png", ".pkm"))   
             
-            if not isAlphaJPG:
-                os.rename(filepath.replace(".png", ".alpha.jpg"), filepath.replace(".png", "_alpha.pkm")) 
-            else:
-                os.rename(filepath.replace(".png", ".alpha.pvr"), filepath.replace(".png", "_alpha.pkm")) 
+            if not isTC4:
+                if not isAlphaJPG:
+                    os.rename(filepath.replace(".png", ".alpha.jpg"), filepath.replace(".png", "_alpha.pkm")) 
+                else:
+                    os.rename(filepath.replace(".png", ".alpha.pvr"), filepath.replace(".png", "_alpha.pkm")) 
 
-            if os.path.exists(filepath.replace(".png", ".alpha.jpg")):
-                os.remove(filepath.replace(".png", ".alpha.jpg"))
-            if os.path.exists(filepath.replace(".png", ".alpha.pvr")):
-                os.remove(filepath.replace(".png", ".alpha.pvr"))            
-        else:    
-            if os.path.exists(filepath.replace(".jpg", ".pkm")):
-                os.remove(filepath.replace(".jpg", ".pkm"))         
-            rgbCMD = """ %s -f ETC1 -p -q etcslow -i %s -o %s """ % (pvrTexToolBin, filepath, filepath.replace(".jpg", ".pvr"))
-            os.system(rgbCMD)
-            os.rename(filepath.replace(".jpg", ".pvr"), filepath.replace(".jpg", ".pkm"))  
+                if os.path.exists(filepath.replace(".png", ".alpha.jpg")):
+                    os.remove(filepath.replace(".png", ".alpha.jpg"))
+                if os.path.exists(filepath.replace(".png", ".alpha.pvr")):
+                    os.remove(filepath.replace(".png", ".alpha.pvr"))  
         
     except Exception:
-        print ("error !!!", filename, "cannot convert.")
-        pass
+        t, v, tb = sys.exc_info()
+        print(t, v)
+        return 2
     finally:
         pass
   
@@ -154,7 +129,6 @@ def work_file_PVR(filename, isAlphaJPG = False):
                 tmpFile.write(b'MNG')
                 
                 rgbname = filepath.replace(".png", ".pkm") 
-                alphaname = filepath.replace(".png", "_alpha.pkm") 
                 
                 statinfo = os.stat(rgbname)
                 fileSize = statinfo.st_size
@@ -164,13 +138,15 @@ def work_file_PVR(filename, isAlphaJPG = False):
                 tmpFile.write(rgbfile.read())
                 rgbfile.close()
                 
-                statinfo = os.stat(alphaname)
-                fileSize = statinfo.st_size
-                
-                tmpFile.write(pack("i", fileSize))
-                alphafile = open(alphaname, "rb")
-                tmpFile.write(alphafile.read())
-                alphafile.close()
+                alphaname = filepath.replace(".png", "_alpha.pkm") 
+                if not isTC4:
+                    statinfo = os.stat(alphaname)
+                    fileSize = statinfo.st_size
+                    
+                    tmpFile.write(pack("i", fileSize))
+                    alphafile = open(alphaname, "rb")
+                    tmpFile.write(alphafile.read())
+                    alphafile.close()
                 
                 # if preAlpha:
                     # tmpFile.write('p')
@@ -184,7 +160,8 @@ def work_file_PVR(filename, isAlphaJPG = False):
                         os.remove(alphaname)
                     
             except Exception:
-                print ("error !!!", filename, "cannot convert.")
+                t, v, tb = sys.exc_info()
+                print(t, v)
                 isSuccess = False
                 pass
             finally: 
@@ -193,48 +170,20 @@ def work_file_PVR(filename, isAlphaJPG = False):
               
         if isSuccess:  
             if isUseGzip:
-                GzipCnt = GzipCnt + 1 
                 gzip_cmd = gzipBin + tmpfilename + " -n -f -9"
                 os.system(gzip_cmd)
                 if os.path.exists(tmpfilename.replace(".tmp", ".png")):
                     os.remove(tmpfilename.replace(".tmp", ".png"))
                 os.rename(tmpfilename + ".gz", tmpfilename.replace(".tmp", ".png"))
+                return 3
             else: 
-                MngCnt = MngCnt + 1
                 if os.path.exists(tmpfilename.replace(".tmp", ".png")):
                     os.remove(tmpfilename.replace(".tmp", ".png"))
                 os.rename(tmpfilename, tmpfilename.replace(".tmp", ".png"))
+                return 5
         else:
-            PassCnt = PassCnt + 1
-            passFiles.append(filename)
-            if os.path.exists(tmpfilename):
-                os.remove(tmpfilename)
-            
-    else:
-        tmpfilename = filepath.replace(".jpg", ".pkm") 
-        
-        if not os.path.exists(tmpfilename):
-            print ("error !!!", filepath, "cannot convert.")
-            PassCnt = PassCnt + 1
-            passFiles.append(filename)
-            return
-        
-        if isUseGzip:
-            GzipCnt = GzipCnt + 1 
-            
-            gzip_cmd = gzipBin + tmpfilename + " -n -f -9"
-            os.system(gzip_cmd)
-            if os.path.exists(tmpfilename.replace(".pkm", ".jpg")):
-                os.remove(tmpfilename.replace(".pkm", ".jpg"))
-            os.rename(tmpfilename + ".gz", tmpfilename.replace(".pkm", ".jpg"))
-        else:
-            PkmCnt = PkmCnt + 1
-            
-            if os.path.exists(tmpfilename.replace(".pkm", ".jpg")):
-                os.remove(tmpfilename.replace(".pkm", ".jpg"))
-            os.rename(tmpfilename, tmpfilename.replace(".pkm", ".jpg"))
- 
+            return 2
 
-def convert(filename):
-    ret = work_file_ETC(filename)
+def convert(filename, isAlphaJPG = False, isTC4 = False):
+    ret = work_file_PVR(filename, isAlphaJPG, isTC4)
     return ret
